apps/stock-data-fetcher/requirements.txt
kafka-python==2.0.2
requests==2.28.1
beautifulsoup4==4.11.1